YH-2064: refactor
