+++ src/_pages/NewLandingPage/ui/NewLandingPage.tsx
@@ -1,21 +1,6 @@
-
-import { QuestionsAndTasksBlock } from "@/widgets/NewLanding/QuestionsAndTasks";
-
-interface NewLandingPage {
-	locale: string;
-}
-
-export const NewLandingPage = ({locale}: NewLandingPage) => {
-
-	return(
-		<>
-			<QuestionsAndTasksBlock locale={locale} />
-		</>
-	)
-}
-
 import { Flex } from '@/shared/ui/Flex';
 import { DifficultySection } from '@/widgets/NewLanding/DifficultySection';
+import { InterviewSection } from '@/widgets/NewLanding/InterviewSection';
 import { PrepareForInterviewSection } from '@/widgets/NewLanding/PrepareForInterviewSection';
 import { TrainerProgressSection } from '@/widgets/NewLanding/TrainerProgressSection';
 
@@ -25,7 +10,7 @@ export const NewLandingPage = () => {
 			<DifficultySection />
 			<PrepareForInterviewSection />
 			<TrainerProgressSection />
+			<InterviewSection />
 		</Flex>
 	);
 };
-
