+++ src/widgets/NewLanding/InterviewSection/index.ts
@@ -1,0 +1,1 @@
+export { InterviewSection } from '@/widgets/NewLanding/InterviewSection/ui/InterviewSection/InterviewSection';
