+++ src/widgets/NewLanding/InterviewSection/model/types.ts
