+++ src/widgets/NewLanding/InterviewSection/ui/InterviewSection/InterviewSection.module.css
@@ -1,0 +1,24 @@
+.cards-list {
+	display: grid;
+	grid-template-columns: repeat(3, 1fr);
+	align-items: flex-end;
+	gap: 20px;
+}
+
+@media screen and (width < 1280px) {
+	.cards-list {
+		max-width: 100%;
+		overflow-x: auto;
+		scrollbar-width: none;
+	}
+}
+
+@media screen and (width < 768px) {
+	.cards-list {
+		grid-template-columns: 1fr;
+		overflow-x: unset;
+		scrollbar-width: unset;
+	}
+}
+
+
