+++ src/widgets/NewLanding/InterviewSection/ui/ListItem/ListItem.module.css
@@ -1,0 +1,143 @@
+a.link {
+	position: relative;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	gap: 8px;
+	padding: 0 24px;
+	height: 48px;
+	border: 1px solid transparent;
+	border-radius: 12px;
+	background-color: var(--button-background-primary);
+	color: var(--text-color-white);
+	text-decoration: none;
+	font-weight: var(--font-weight-400);
+	font-size: var(--font-size-p-m);
+	transition: background-color 0.2s ease;
+	cursor: pointer;
+}
+
+a.link:hover {
+	background-color: var(--button-background-primary-hover);
+	color: var(--text-color-white);
+	text-decoration: none;
+}
+
+.card {
+	position: relative;
+	padding: 20px;
+	height: 270px;
+	overflow: hidden;
+	box-shadow: 0 4px 10px 0 #6A63761A;
+	border-radius: 20px;
+	background-color: var(--background-block);
+	transition: height 0.3s ease, transform 0.3s ease, width 0.3s ease, opacity 0.3s ease;
+	cursor: pointer;
+}
+
+.card:hover {
+	z-index: 2;
+	height: 396px;
+}
+
+.card-img {
+	position: absolute;
+	top: 50px;
+	transition: opacity 0.3s ease;
+	opacity: 0;
+}
+
+.indicator{
+	margin-top: 9px;
+	opacity: 1;
+}
+
+.card:hover .card-img {
+	opacity: 1;
+}
+
+.card:hover .indicator {
+	opacity: 0;
+}
+
+.card-content {
+	position: relative;
+	height: 100%;
+}
+
+.text {
+	transition: margin-top 0.3s ease-in-out, margin-bottom 0.3s ease-in-out;
+}
+
+.img-quest {
+	left: 50%;
+	width: calc(100% - 40px);
+	transform: translateX(-50%);
+}
+
+.img-task {
+	left: 20%;
+	width: 100%;
+}
+
+.img-collect {
+	left: 10%;
+	width: 100%;
+}
+
+.card-img img {
+	display: block;
+	width: 100%;
+	height: auto;
+}
+
+@media (width < 1280px) {
+	p.card-title {
+		font-size: var(--font-size-h-xs);
+	}
+}
+
+@media (width < 1024px){
+	.card {
+		width: 320px;
+		height: 293px;
+	}
+
+	.card:hover {
+		width: 344px;
+		height: 358px;
+	}
+}
+
+@media (width < 850px){
+	.card {
+		width: 253px;
+	}
+}
+
+@media (width < 768px){
+	.card {
+		flex-shrink: 1;
+		padding: 16px;
+		width: 100%;
+		height: 358px;
+	}
+	
+	.indicator {
+		opacity: 0;
+	}
+
+	.text {
+		margin-bottom: 16px;
+	}
+
+	.card:hover {
+		width: 100%;
+		height: 358px;
+	}
+
+	.card-img {
+		opacity: 1;
+		top: 20px;
+	}
+}
