--- src/widgets/NewLanding/QuestionsAndTasks/index.ts
@@ -1,1 +1,0 @@
-export { QuestionsAndTasksBlock } from './ui/QuestionsAndTasksBlock/QuestionsAndTasksBlock';
